+++ libs/grid-angular/features/master-detail/src/index.ts
@@ -1,0 +1,21 @@
+/**
+ * Master-detail feature for @toolbox-web/grid-angular
+ *
+ * Import this module to enable the `masterDetail` input on Grid directive.
+ *
+ * @example
+ * ```typescript
+ * import '@toolbox-web/grid-angular/features/master-detail';
+ *
+ * <tbw-grid [masterDetail]="{ detailRenderer: myRenderer }" />
+ * ```
+ *
+ * @packageDocumentation
+ */
+
+import { MasterDetailPlugin } from '@toolbox-web/grid/plugins/master-detail';
+import { registerFeature } from '@toolbox-web/grid-angular';
+
+registerFeature('masterDetail', (config) => {
+  return new MasterDetailPlugin(config ?? undefined);
+});